+++ inc/Acf/LocalJsonCacheSettings.php
@@ -3,6 +3,7 @@
 namespace Airfleet\Framework\Acf;
 
 use Airfleet\Framework\Features\Feature;
+use Airfleet\Framework\Helpers\Environment;
 
 class LocalJsonCacheSettings implements Feature {
 	protected static ?LocalJsonCacheSettings $instance = null;
@@ -11,8 +12,9 @@ class LocalJsonCacheSettings implements Feature {
 	protected int $expiration;
 
 	public function initialize(): void {
-		$this->enabled = apply_filters( 'airfleet/framework/acf/local_json_cache/enabled', true );
-		$this->expiration = apply_filters( 'airfleet/framework/acf/local_json_cache/expiration', DAY_IN_SECONDS );
+		$default_enabled = Environment::is_local() ? false : true;
+		$this->enabled = apply_filters( 'airfleet/acf/local_json_cache/enabled', $default_enabled );
+		$this->expiration = apply_filters( 'airfleet/acf/local_json_cache/expiration', DAY_IN_SECONDS );
 	}
 
 	protected static function instance(): LocalJsonCacheSettings {