+++ CHANGELOG.md
@@ -7,6 +7,34 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Added Environment helpers (`Helpers\Environment`) - determine environment (local, staging, production)
+- Added `CacheManager` utility class for working with cache
+- Added `Cacheable` trait to easily access global `CacheManager` instance
+- Added filter `airfleet/cache/enabled` to enable/disable cache
+- Added filter `airfleet/cache/expiration` to set cache duration in seconds (0 to not expire)
+- Added filter `airfleet/cache/cache_dir_checks` to enable/disable caching directory modification checks
+- Added filter `airfleet/cache/expiration_dir_check` to set cache duration for directory modification checks in seconds (0 to not expire)
+- Added filter `airfleet/cache/manager` to set cache manager
+- Added action `airfleet/cache/flush` to clear the cache
+- Added action `airfleet/cache/hit` that is triggered whenever cache is hit, but only if WP_DEBUG is enabled
+- Added action `airfleet/cache/miss` that is triggered whenever there is a cache miss, but only if WP_DEBUG is enabled
+- Added class `Benchmark` to setup benchmarking feature. When enabled, an HTML comment will be printed at the end of the page with the amount of time in milliseconds it took to run certain features. The comment is only displayed for authenticated users.
+- Added filter `airfleet/benchmark/enabled` to enable/disable performance benchmark
+- Added filter `airfleet/benchmark/log_to_error_log` to enable/disable logging benchmark results to the error log. This is in addition to the HTML comment and dependent on benchmarking being enabled
+- Added action `airfleet/benchmark/start` to start tracking a workload benchmark (must pass ID)
+- Added action `airfleet/benchmark/stop` to stop tracking a workload benchmark (must pass ID)
+- Added utility classes `Timer` and `NestedTimerRegistry` used for benchmarking
+
+### Changed
+
+- `LocalJsonLoadFromCache`: Refactored to use `Cacheable` and improve cache invalidation when ACF Local JSON directory contents change
+- `LocalJsonCacheSettings`: Automatically set default value for filter `airfleet/cache/enabled` based on environment (enabled for all except local)
+- Renamed filter `airfleet/framework/acf/local_json_cache/enabled` to `airfleet/acf/local_json_cache/enabled`
+- Renamed filter `airfleet/framework/acf/local_json_cache/expiration` to `airfleet/acf/local_json_cache/expiration`
+- Renamed action `airfleet/framework/acf/local_json_cache/invalidate` to `airfleet/acf/local_json_cache/invalidate`
+
 ## [1.17.1] - 2025-09-15
 
 ### Fixed
@@ -271,5 +299,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [1.15.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/1.15.0
 [1.16.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/1.16.0
 [1.17.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/1.17.0
-
 [1.17.1]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/1.17.1